# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Venue management API for creating venues, branches, and resources
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseDetailsItem = { [key: string]: unknown };

export interface ErrorResponse {
  error: string;
  details?: ErrorResponseDetailsItem[];
}

export interface CreateVenueRequest {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /** @maxLength 2000 */
  description?: string;
  /** @maxLength 300 */
  address?: string;
  /** @maxLength 120 */
  city?: string;
  /** @maxLength 120 */
  country?: string;
  /** @maxLength 80 */
  timezone?: string;
  /** @maxLength 320 */
  contactEmail?: string;
  /** @maxLength 40 */
  contactPhone?: string;
  /** @maxLength 500 */
  website?: string;
}

export type Venue = CreateVenueRequest & {
  id: string;
  createdAt: string;
  updatedAt: string;
};

export interface CreateBranchRequest {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  code?: string;
  /** @maxLength 2000 */
  description?: string;
  /** @maxLength 300 */
  address?: string;
  /** @maxLength 120 */
  city?: string;
  /** @maxLength 120 */
  country?: string;
  /** @maxLength 80 */
  timezone?: string;
  /** @maxLength 40 */
  contactPhone?: string;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude?: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude?: number;
}

export type Branch = CreateBranchRequest & {
  id: string;
  venueId: string;
  createdAt: string;
  updatedAt: string;
};

export type CreateResourceRequestMetadata = { [key: string]: unknown };

export interface CreateResourceRequest {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  type: string;
  /** @maxLength 2000 */
  description?: string;
  /**
     * @minimum 1
     * @maximum 1000000
     */
  capacity?: number;
  isActive?: boolean;
  metadata?: CreateResourceRequestMetadata;
}

export type Resource = CreateResourceRequest & {
  id: string;
  branchId: string;
  createdAt: string;
  updatedAt: string;
};

/**
 * Request validation failed
 */
export type ValidationErrorResponse = ErrorResponse;

/**
 * Parent entity was not found
 */
export type NotFoundResponse = ErrorResponse;