# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Venue management API for creating venues, branches, and resources
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: venues
    description: Venue operations
  - name: branches
    description: Branch operations
  - name: resources
    description: Resource operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /venues:
    post:
      operationId: createVenue
      tags: [venues]
      summary: Create a venue
      description: Creates a venue that can contain one or more branches.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueRequest"
      responses:
        "201":
          description: Venue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          $ref: "#/components/responses/ValidationError"
  /venues/{venueId}/branches:
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      description: Creates a branch under an existing venue.
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchRequest"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /branches/{branchId}/resources:
    post:
      operationId: createResource
      tags: [resources]
      summary: Create a resource
      description: Creates a bookable or manageable resource under an existing branch.
      parameters:
        - $ref: "#/components/parameters/BranchId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceRequest"
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    VenueId:
      name: venueId
      in: path
      required: true
      description: Venue identifier
      schema:
        type: string
        format: uuid
    BranchId:
      name: branchId
      in: path
      required: true
      description: Branch identifier
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Parent entity was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - error
    CreateVenueRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 2000
        address:
          type: string
          maxLength: 300
        city:
          type: string
          maxLength: 120
        country:
          type: string
          maxLength: 120
        timezone:
          type: string
          maxLength: 80
        contactEmail:
          type: string
          format: email
          maxLength: 320
        contactPhone:
          type: string
          maxLength: 40
        website:
          type: string
          format: uri
          maxLength: 500
      required:
        - name
    Venue:
      allOf:
        - $ref: "#/components/schemas/CreateVenueRequest"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required:
            - id
            - createdAt
            - updatedAt
    CreateBranchRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        code:
          type: string
          minLength: 1
          maxLength: 40
        description:
          type: string
          maxLength: 2000
        address:
          type: string
          maxLength: 300
        city:
          type: string
          maxLength: 120
        country:
          type: string
          maxLength: 120
        timezone:
          type: string
          maxLength: 80
        contactPhone:
          type: string
          maxLength: 40
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
      required:
        - name
    Branch:
      allOf:
        - $ref: "#/components/schemas/CreateBranchRequest"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            venueId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required:
            - id
            - venueId
            - createdAt
            - updatedAt
    CreateResourceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        type:
          type: string
          minLength: 1
          maxLength: 80
        description:
          type: string
          maxLength: 2000
        capacity:
          type: integer
          minimum: 1
          maximum: 1000000
        isActive:
          type: boolean
          default: true
        metadata:
          type: object
          additionalProperties: true
      required:
        - name
        - type
    Resource:
      allOf:
        - $ref: "#/components/schemas/CreateResourceRequest"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            branchId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required:
            - id
            - branchId
            - createdAt
            - updatedAt