# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Venue management API for creating venues, branches, and resources
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBranchRequest } from './createBranchRequest';

export type Branch = CreateBranchRequest & {
  id: string;
  venueId: string;
  createdAt: Date;
  updatedAt: Date;
};