# lib/api-zod/src/generated/types/createBranchRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Venue management API for creating venues, branches, and resources
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBranchRequest {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  code?: string;
  /** @maxLength 2000 */
  description?: string;
  /** @maxLength 300 */
  address?: string;
  /** @maxLength 120 */
  city?: string;
  /** @maxLength 120 */
  country?: string;
  /** @maxLength 80 */
  timezone?: string;
  /** @maxLength 40 */
  contactPhone?: string;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude?: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude?: number;
}