# lib/api-zod/src/generated/types/createResourceRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Venue management API for creating venues, branches, and resources
 * OpenAPI spec version: 0.1.0
 */
import type { CreateResourceRequestMetadata } from './createResourceRequestMetadata';

export interface CreateResourceRequest {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  type: string;
  /** @maxLength 2000 */
  description?: string;
  /**
     * @minimum 1
     * @maximum 1000000
     */
  capacity?: number;
  isActive?: boolean;
  metadata?: CreateResourceRequestMetadata;
}