# lib/api-zod/src/generated/types/createVenueRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Venue management API for creating venues, branches, and resources
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVenueRequest {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /** @maxLength 2000 */
  description?: string;
  /** @maxLength 300 */
  address?: string;
  /** @maxLength 120 */
  city?: string;
  /** @maxLength 120 */
  country?: string;
  /** @maxLength 80 */
  timezone?: string;
  /** @maxLength 320 */
  contactEmail?: string;
  /** @maxLength 40 */
  contactPhone?: string;
  /** @maxLength 500 */
  website?: string;
}