# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Venue management API for creating venues, branches, and resources
 * OpenAPI spec version: 0.1.0
 */
import type { ErrorResponseDetailsItem } from './errorResponseDetailsItem';

export interface ErrorResponse {
  error: string;
  details?: ErrorResponseDetailsItem[];
}