# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Venue management API for creating venues, branches, and resources
 * OpenAPI spec version: 0.1.0
 */
import type { CreateResourceRequest } from './createResourceRequest';

export type Resource = CreateResourceRequest & {
  id: string;
  branchId: string;
  createdAt: Date;
  updatedAt: Date;
};