# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Venue management API for creating venues, branches, and resources
 * OpenAPI spec version: 0.1.0
 */
import type { CreateVenueRequest } from './createVenueRequest';

export type Venue = CreateVenueRequest & {
  id: string;
  createdAt: Date;
  updatedAt: Date;
};